+++ library/ui/src/main/java/com/google/android/exoplayer2/ui/StyledPlayerControlView.java
@@ -490,7 +490,11 @@ public class StyledPlayerControlView extends FrameLayout {
       TypedArray a =
           context
               .getTheme()
-              .obtainStyledAttributes(playbackAttrs, R.styleable.StyledPlayerControlView, 0, 0);
+              .obtainStyledAttributes(
+                  playbackAttrs,
+                  R.styleable.StyledPlayerControlView,
+                  defStyleAttr,
+                  /* defStyleRes= */ 0);
       try {
         controllerLayoutId =
             a.getResourceId(
@@ -575,8 +579,8 @@ public class StyledPlayerControlView extends FrameLayout {
     if (customTimeBar != null) {
       timeBar = customTimeBar;
     } else if (timeBarPlaceholder != null) {
-      // Propagate attrs as timebarAttrs so that DefaultTimeBar's custom attributes are transferred,
-      // but standard attributes (e.g. background) are not.
+      // Propagate playbackAttrs as timebarAttrs so that DefaultTimeBar's custom attributes are
+      // transferred, but standard attributes (e.g. background) are not.
       DefaultTimeBar defaultTimeBar =
           new DefaultTimeBar(context, null, 0, playbackAttrs, R.style.ExoStyledControls_TimeBar);
       defaultTimeBar.setId(R.id.exo_progress);